Support OpenSSL 3.0.8
The multi-platform compilation and adaptation of OpenSSL 3.0.8 was completed, which mainly upgraded the lua-openssl repository. Luasocket and luasec were not modified as they were not affected and passed compatibility testing. In addition to clarifying the source code of the plugin itself, adaptations and changes were made to the plugin template, and support was added for Apple arm64 Simulator and xcframeworks, providing reference for handling future plugin issues.